Make table tests easier to read.
